Vastly improved OverlayScreenComponent
Now you can hack screens in front of your managed screens!
